Add files via upload
This program has both encode and decode functionality.
The algorithm for encoding message has been changed. 
This program uses Bit-wise And operation and only works for letters between A-Z (Uppercase).
XOR is applied to each character of the message and mask is added. (first 5 bits from right to left is same for A and a)